refactor(Settings*): move generic settings components to shared/
