+++ src/pages/options/Donate.tsx
@@ -4,7 +4,7 @@ import Typography from '@mui/material/Typography';
 import { usePageTitle } from '@root/src/shared/contexts/PageTitleContext';
 import Rolod0xText from '@src/components/Rolod0xText';
 
-import SettingsPageHeader from './SettingsPageHeader';
+import SettingsPageHeader from './shared/SettingsPageHeader';
 import StyledCode from './shared/StyledCode';
 
 export default function Donate() {
