+++ src/pages/options/shared/SettingsAccordionSection.tsx
